pyproject.toml
[project]
name = "tuffysearch-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastapi[all]>=0.115.12",
    "langchain-community>=0.3.24",
    "langchain-huggingface>=0.2.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "slowapi>=0.1.9",
]
